server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.30",
    "celery>=5.5.3",
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "pmtiles>=3.4.1",
    "pre-commit>=4.3.0",
    "psycopg[binary]>=3.2.10",
    "redis>=6.4.0",
    "ruff>=0.13.0",
]
